# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tenant Payment & Communication API",
    "description": "REST API for property management, lease handling, payment processing, and tenant communication. Includes user authentication, property management, payment tracking, maintenance requests, messaging, and AI-powered features.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    },
    {
      "url": "https://api.example.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/maintenance": {
      "get": {
        "tags": ["Maintenance"],
        "summary": "List maintenance requests",
        "description": "Get maintenance requests with filters",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "enum": ["pending", "in-progress", "completed", "cancelled"] }
          },
          {
            "name": "priority",
            "in": "query",
            "schema": { "type": "string", "enum": ["low", "medium", "high"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance requests retrieved successfully"
          }
        }
      },
      "post": {
        "tags": ["Maintenance"],
        "summary": "Create maintenance request",
        "description": "Create a new maintenance request (tenant only)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["unitId", "title", "description", "category"],
                "properties": {
                  "unitId": { "type": "string", "format": "uuid" },
                  "title": { "type": "string", "example": "Leaky faucet" },
                  "description": { "type": "string" },
                  "category": { "type": "string", "enum": ["plumbing", "electrical", "structural", "appliance", "other"] },
                  "priority": { "type": "string", "enum": ["low", "medium", "high"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Maintenance request created successfully"
          }
        }
      }
    },
...
    "/api/messages/conversations": {
      "get": {
        "tags": ["Messages"],
        "summary": "List conversations",
        "description": "Get all conversations for the user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Conversations retrieved successfully"
          }
        }
      },
      "post": {
        "tags": ["Messages"],
        "summary": "Create conversation",
        "description": "Start a new conversation",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["participantId"],
                "properties": {
                  "participantId": { "type": "string", "format": "uuid" },
                  "subject": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Conversation created successfully"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token received from login endpoint"
      }
    },
...
  }
}